Rename STACK_POINTER_SIZE to STACK_ENTRY_SIZE
Stack pointer is analogous to PC; it points at the current element of the stack. It's size should be enough to be able to point to every valid stack location.

Stack pointer size doesn't determine the increment or decrement when items are pushed onto or popped off from the stack. This is the role of the entry size.